% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_url}
\alias{api_url}
\title{URL link for agviewer.com/api/v1
\itemize{
\item \code{ url_api } : url
}}
\format{
A \code{ character } with url
}
\usage{
api_url
}
\description{
URL link for agviewer.com/api/v1
\itemize{
\item \code{ url_api } : url
}
}
\keyword{datasets}
