% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estaciones_garces}
\alias{estaciones_garces}
\title{Estaciones agromet}
\format{
A \code{data frame} with 199  observations and 5 variables.
}
\usage{
estaciones_garces
}
\description{
Estaciones agromet
}
\section{Variables}{


\itemize{#'
\item \code{ nombre } : station name.
\item \code{ serial } : serial identification.
\item \code{ lat } : decimal degree for latitude coordinate.
\item \code{ lon } : decimal degree for longitud coordinate.
\item \code{ tipo } : if it is for soil moisture or not.
}
}

\keyword{datasets}
