% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataAGV.R
\name{getdataAGV}
\alias{getdataAGV}
\title{Get raw list-column data from agviewer.com/api/v2}
\usage{
getdataAGV(station_id, time_span)
}
\arguments{
\item{station_id}{serial number of the station}

\item{time_span}{time_span character vector of dates (\%Y-\%m-\%d) having c(start_date,end_date)}
}
\value{

}
\description{
Get raw list-column data from agviewer.com/api/v2
}
