% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataAGV_clima.R
\name{getDataAGV_clima}
\alias{getDataAGV_clima}
\title{Get climatic data from agviewer.com/api/v1}
\usage{
getDataAGV_clima(
  station_id = "z6-14410",
  var = "Temperature",
  time_span = c("2022-07-20", "2022-07-25")
)
}
\arguments{
\item{station_id}{serial number of the station}

\item{var}{pattern of the variable name to be search via regular expression}

\item{time_span}{character vector of dates (\%Y-\%m-\%d) having c(start_date,end_date)}
}
\description{
Get climatic data from agviewer.com/api/v1
}
\examples{
data1 <- getDataAGV_clima(station_id ='z6-12948',var = 'Temperature', time_span = c("2022-07-27",'2022-07-28')) 
data2 <- getDataAGV_clima(station_id ='0020500D',var = 'Temperature', time_span = c("2022-07-27",'2022-07-28'))
data3 <- getDataAGV_clima(station_id ='00203581',var = 'Temperature', time_span = c("2022-07-27",'2022-07-28')) 
}
