% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdataAGV_sm.R
\name{getdataAGV_sm}
\alias{getdataAGV_sm}
\title{Get soil moisture data from agviewer.com/api/v1}
\usage{
getdataAGV_sm(
  station_id = "z6-08674",
  time_span = c(Sys.Date() - 1, Sys.Date())
)
}
\arguments{
\item{station_id}{serial number of the station}

\item{time_span}{character vector of dates (\%Y-\%m-\%d) having c(start_date,end_date)}
}
\description{
Get soil moisture data from agviewer.com/api/v1
}
